+++ src/pointer/any.rs
@@ -38,7 +38,7 @@ impl<'value> NP_Value<'value> for NP_Any {
     fn get_size<M: NP_Memory>(_cursor: &NP_Cursor, _memory: &M) -> Result<usize, NP_Error> {
         Ok(0)
     }
-    fn do_compact<M: NP_Memory>(from_cursor: NP_Cursor, from_memory: &'value M, to_cursor: NP_Cursor, to_memory: &'value M) -> Result<NP_Cursor, NP_Error> where Self: 'value + Sized {
+    fn do_compact<M: NP_Memory, M2: NP_Memory>(from_cursor: NP_Cursor, from_memory: &'value M, to_cursor: NP_Cursor, to_memory: &'value M2) -> Result<NP_Cursor, NP_Error> where Self: 'value + Sized {
         Err(NP_Error::new("Cannot compact an ANY field!"))
     }
     fn from_json_to_schema(mut schema: Vec<NP_Parsed_Schema>, _json_schema: &Box<NP_JSON>) -> Result<(bool, Vec<u8>, Vec<NP_Parsed_Schema>), NP_Error> {