+++ src/pointer/string.rs
@@ -326,8 +326,7 @@ impl<'value> NP_Value<'value> for &'value str {
     
         let str_size = bytes.len() as usize;
     
-        let write_bytes = memory.write_bytes();
-    
+        let mut write_bytes = memory.write_bytes();    
 
         if size > 0 {
             // fixed size bytes
@@ -345,6 +344,7 @@ impl<'value> NP_Value<'value> for &'value str {
             }
 
             let addr = c_value.get_addr_value() as usize;
+            write_bytes = memory.write_bytes();
     
             for x in 0..(size as usize) {
                 if x < bytes.len() {